requirements.txt
opencv-python==4.8.1.78
tensorflow==2.13.0
numpy==1.24.3
pillow==10.0.1
scikit-learn==1.3.0
matplotlib==3.7.2
face-recognition==1.3.0
dlib==19.24.2
PyQt5==5.15.9
threading-timer==0.1.0
